Add ForgotPassword component with email verification
Created a ForgotPassword component that allows users to reset their password by entering their email address.
Integrated Firebase's sendPasswordResetEmail to send a verification code to the user's email.
Displays a message confirming the email was sent or an error if the email couldn't be sent.
Added a link to redirect users back to the login page after requesting a password reset.